+++ packages/uba/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- Added captionSingular for all *.meta and *.meta.lang
+
 ### Changed
 
 ### Deprecated
@@ -14,6 +14,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.23.10] - 2022-08-04
+### Added
+- Added captionSingular for all *.meta and *.meta.lang
+- `uba_group` form - multiselect for users in group
+
+### Removed
+ - for attributes of type `Boolean` removed `allowNull: false` and `defaultValue` - not needed anymore
+
+### Fixed
+- `uba_group` form - fix unnecessary table refreshes
 
 ## [5.23.9] - 2022-07-28
 ## [5.23.8] - 2022-07-26